Listコンポーネント、Formコンポーネントをそれぞれ別ファイルに作成し動作確認
